BE-641: Split the select statement module and add quantifier builder sugar
